TST: Move mocks in service.py into mocks.py
